+++ NEWS.rst
@@ -42,6 +42,7 @@ New Features
 * `hy.repr` now supports several more standard types.
 * the attribute access macro `.` now accepts method calls.
 * `as_model` checks for self-references in its argument.
+* New function `hy.model_patterns.keepsym`
 
 .. _Hyrule: https://github.com/hylang/hyrule